First compiling version

+++ core/kserver_session.cpp
@@ -76,7 +76,7 @@ int Session<TCP>::read_command(Command& cmd)
 }
 
 template<>
-int Session<TCP>::rcv_n_bytes(char *buffer, uint32_t n_bytes)
+int Session<TCP>::rcv_n_bytes(char *buffer, uint64_t n_bytes)
 {
     int bytes_rcv = 0;
     uint32_t bytes_read = 0;
